refactor pre-server logic
