+++ src/client/index.ts
@@ -1,0 +1,60 @@
+import {ExtensionContext, LanguageClient, LanguageClientOptions, ServerOptions, TransportKind, workspace} from "coc.nvim";
+import * as path from 'path';
+
+export default function setupLwcServerClient(context: ExtensionContext): LanguageClient {
+  // Setup the language server
+  const serverModule = context.asAbsolutePath(
+    path.join(
+      'node_modules',
+      '@salesforce',
+      'lwc-language-server',
+      'lib',
+      'server.js'
+    )
+  );
+
+
+  const debugOptions = { execArgv: ['--nolazy', '--inspect=6009'] };
+  const serverOptions: ServerOptions = {
+    run: {
+      module: serverModule,
+      transport: TransportKind.ipc
+    },
+    debug: {
+      module: serverModule,
+      transport: TransportKind.ipc,
+      options: debugOptions
+    }
+  };
+
+  const clientOptions: LanguageClientOptions = {
+    documentSelector: [
+      { language: 'html', scheme: 'file' },
+      { language: 'javascript', scheme: 'file' }
+    ],
+    synchronize: {
+      fileEvents: [
+        workspace.createFileSystemWatcher('**/*.resource'),
+        workspace.createFileSystemWatcher(
+          '**/labels/CustomLabels.labels-meta.xml'
+        ),
+        workspace.createFileSystemWatcher(
+          '**/staticresources/*.resource-meta.xml'
+        ),
+        workspace.createFileSystemWatcher('**/contentassets/*.asset-meta.xml'),
+        workspace.createFileSystemWatcher('**/lwc/*/*.js'),
+        workspace.createFileSystemWatcher('**/modules/*/*/*.js'),
+        // need to watch for directory deletions as no events are created for contents or deleted directories
+        workspace.createFileSystemWatcher('**/', false, true, false)
+      ]
+    }
+  };
+
+  // Create the language client and start the client.
+  return new LanguageClient(
+    'lwcLanguageServer',
+    'LWC Language Server',
+    serverOptions,
+    clientOptions
+  );
+}
